使用异步化的全新方式改进数据库对象模型的设计，并且支持完整的事务隔离级别

+++ lealone-sql/src/main/java/org/lealone/sql/ddl/DropDatabase.java
@@ -19,10 +19,10 @@ package org.lealone.sql.ddl;
 
 import org.lealone.common.exceptions.DbException;
 import org.lealone.db.Database;
-import org.lealone.db.DbObjectType;
 import org.lealone.db.LealoneDatabase;
 import org.lealone.db.api.ErrorCode;
 import org.lealone.db.session.ServerSession;
+import org.lealone.db.table.LockTable;
 import org.lealone.sql.SQLStatement;
 
 /**
@@ -58,23 +58,25 @@ public class DropDatabase extends DatabaseStatement {
             throw DbException.get(ErrorCode.CANNOT_DROP_LEALONE_DATABASE);
         }
         LealoneDatabase lealoneDB = LealoneDatabase.getInstance();
-        synchronized (lealoneDB.getLock(DbObjectType.DATABASE)) {
-            Database db = lealoneDB.getDatabase(dbName);
-            if (db == null) {
-                if (!ifExists)
-                    throw DbException.get(ErrorCode.DATABASE_NOT_FOUND_1, dbName);
-            } else {
-                lealoneDB.removeDatabaseObject(session, db);
-                if (isTargetNode(db)) {
-                    // Lealone不同于H2数据库，在H2的一个数据库中可以访问另一个数据库的对象，而Lealone不允许，
-                    // 所以在H2中需要一个对象一个对象地删除，这样其他数据库中的对象对他们的引用才能解除，
-                    // 而Lealone只要在LealoneDatabase中删除对当前数据库的引用然后删除底层的文件即可。
-                    if (deleteFiles) {
-                        db.setDeleteFilesOnDisconnect(true);
-                    }
-                    if (db.getSessionCount() == 0) {
-                        db.drop();
-                    }
+        LockTable lockTable = lealoneDB.tryExclusiveDatabaseLock(session);
+        if (lockTable == null)
+            return -1;
+
+        Database db = lealoneDB.getDatabase(dbName);
+        if (db == null) {
+            if (!ifExists)
+                throw DbException.get(ErrorCode.DATABASE_NOT_FOUND_1, dbName);
+        } else {
+            lealoneDB.removeDatabaseObject(session, db);
+            if (isTargetNode(db)) {
+                // Lealone不同于H2数据库，在H2的一个数据库中可以访问另一个数据库的对象，而Lealone不允许，
+                // 所以在H2中需要一个对象一个对象地删除，这样其他数据库中的对象对他们的引用才能解除，
+                // 而Lealone只要在LealoneDatabase中删除对当前数据库的引用然后删除底层的文件即可。
+                if (deleteFiles) {
+                    db.setDeleteFilesOnDisconnect(true);
+                }
+                if (db.getSessionCount() == 0) {
+                    db.drop();
                 }
             }
         }
